src: Implement new node access hint system
v4l2loopback has a new patch queued up for getting hints by
the kernel module to decide on whether a device is supposed
to be fed buffers or not. This will be used for permission
dialogs in the long run, but for now feeding frames is
the main goal of this change.